Adding error logging and moving some tests to the main package
